Take the weaviate class name from environment variable
