+++ airflow/Dockerfile
@@ -1,5 +1,5 @@
 # syntax=quay.io/astronomer/airflow-extensions:latest
 
-FROM quay.io/astronomer/astro-runtime:9.3.0-base
+FROM quay.io/astronomer/astro-runtime:9.4.0-base
 
 COPY include/airflow_provider_weaviate-0.0.1-py3-none-any.whl /tmp
