+++ airflow/dags/monitor/monitor.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import json
 import logging
 import os
@@ -6,7 +8,6 @@ from datetime import datetime
 
 import firebase_admin
 import requests
-from monitor.external_api import APIMonitoring
 from weaviate_provider.hooks.weaviate import WeaviateHook
 
 from airflow.decorators import dag, task
@@ -32,6 +33,41 @@ ASK_ASTRO_API_BASE_URL = os.environ.get("ASK_ASTRO_API_BASE_URL", None)
 # request ids to monitor from Ask-Astro database.
 ASK_ASTRO_REQUEST_ID_1 = os.environ.get("ASK_ASTRO_REQUEST_ID_1", "05d8882e-56ac-11ee-a818-4200a9fe0102")
 ASK_ASTRO_REQUEST_ID_2 = os.environ.get("ASK_ASTRO_REQUEST_ID_2", "f2d6524c-56ab-11ee-a818-4200a9fe0102")
+
+
+class APIMonitoring:
+    """
+    A class to test API endpoints using GET or POST requests and ensure that they return HTTP 200 status codes.
+    """
+
+    def __init__(self, base_url: str) -> None:
+        self.base_url = base_url
+
+    def test_endpoint(
+        self, endpoint: str, method: str = "GET", data: dict | None = None, headers: dict | None = None
+    ) -> int:
+        """
+        Test an endpoint with the specified method, data, and headers.
+
+        :param endpoint: The endpoint to test.
+        :param method: The HTTP method to use. Defaults to 'GET'.
+        :param data: The data to send in the request. Defaults to None.
+        :param headers: The headers to send in the request. Defaults to None.
+        """
+        try:
+            url = f"{self.base_url}{endpoint}"
+            if method.upper() == "GET":
+                response = requests.get(url, headers=headers)
+            elif method.upper() == "POST":
+                response = requests.post(url, json=data, headers=headers)
+            else:
+                logger.error(f"Unsupported method: {method}")
+                raise Exception(f"Unsupported method: {method}")
+            response.raise_for_status()
+            logger.info(f"Success: {method} {endpoint} returned status code {response.status_code}")
+
+        except requests.exceptions.HTTPError as err:
+            logger.error(f"Failed: {method} {endpoint} returned status code {err.response.status_code}")
 
 
 @task(trigger_rule=TriggerRule.ALL_DONE)
@@ -127,11 +163,12 @@ def monitor_apis():
     This task will test each API endpoint defined in the endpoints list and will print the status code for each.
     If any endpoint does not return a 200 status, an exception will be raised.
     """
-    headers = {"accept": "*/*", "Content-Type": "application/json"}
     if not ASK_ASTRO_API_BASE_URL:
         raise ValueError("base_url cannot be empty.")
+
     api_monitor = APIMonitoring(base_url=ASK_ASTRO_API_BASE_URL)
 
+    headers = {"accept": "*/*", "Content-Type": "application/json"}
     request_body = {"prompt": "Example prompt"}
     feedback_body = {"positive": True}
 
@@ -150,8 +187,8 @@ def monitor_apis():
     for endpoint_info in endpoints_to_monitor:
         try:
             # Unpack the tuple with a default value for 'body'
-            endpoint, method, headers, body = (*endpoint_info, None)[:4]
-            status_code = api_monitor.test_endpoint(endpoint, method, headers, body)
+            endpoint, method, headers, data = (*endpoint_info, None)[:4]
+            status_code = api_monitor.test_endpoint(endpoint=endpoint, method=method, headers=headers, data=data)
             logger.info(f"Endpoint {endpoint} returned status code {status_code}")
         except requests.HTTPError as e:
             logger.error(f"Endpoint {endpoint_info[0]} failed with status code: {e.response.status_code}")
